cleaning up constants
