+++ lib/constants.ts
@@ -1,3 +1,6 @@
+import { darkTheme, Theme } from "@rainbow-me/rainbowkit";
+import { merge } from "lodash";
+
 // JWT stuff
 export const JWT_SECRET_KEY = process.env.INFINITY_KEYS_SECRET;
 export const IK_ID_COOKIE = "ik-id";
@@ -17,3 +20,22 @@ export const PAGINATION_COUNTS = [16, 32, 64];
 
 // Cloudinary
 export const CLOUDINARY_CLOUD_NAME = "infinity-keys";
+
+// Rainbow Kit Styling
+export const IKTheme = merge(darkTheme(), {
+  colors: {
+    accentColor: "#3FCCBB",
+    connectButtonBackground: "#354161",
+    modalBackground: "#101D42",
+    modalBorder: "rgba(255,255,255,.2)",
+    menuItemBackground: "#101D42",
+  },
+  fonts: {
+    body: "Poppins, sans-serif",
+  },
+  radii: {
+    connectButton: "4px",
+    modal: "8px",
+    modalMobile: "8px",
+  },
+} as Theme);
