+++ pages/api/minter/verify.ts
@@ -2,18 +2,10 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { ethers } from "ethers";
 import castArray from "lodash/castArray";
-import {
-  AVAX_CHAIN_ID,
-  CONTRACT_ADDRESS_AVAX,
-  CONTRACT_ADDRESS_ETH,
-  CONTRACT_ADDRESS_POLYGON,
-  ETH_CHAIN_ID,
-  POLYGON_CHAIN_ID,
-} from "@lib/walletConstants";
 import { IK_ID_COOKIE } from "@lib/constants";
 import { gqlApiSdk } from "@lib/server";
 import { jwtHasClaim } from "@lib/jwt";
-import { contractAddressLookup } from "@lib/contracts";
+import { contractAddressLookup } from "@lib/walletConstants";
 
 const privateKey = process.env.PRIVATE_KEY_VERIFY;
 const secret = process.env.MINT_SECRET_VERIFY;
